refactor: split CanvasEditor side panels
