+++ src/components/CanvasEditor.jsx
@@ -48,6 +48,9 @@ import { jsPDF } from "jspdf";
 import TemplateLayout from "../Pages/addTemplateLayout";
 import PrintSettings from "./PrintSettings";
 import FrameSection from "./FrameSection";
+import FiltersSection from "./FiltersSection";
+import SelectedObjectSection from "./SelectedObjectSection";
+import ImageToolsSection from "./ImageToolsSection";
 import { buildClipShape, buildOverlayShape, moveOverlayAboveImage, applyMaskAndFrame, removeMaskAndFrame } from "../utils/shapeUtils";
 import { PRESET_SIZES, mmToPx, pxToMm, drawCropMarks, drawRegistrationMark } from "../utils/printUtils";
 
@@ -1470,65 +1473,20 @@ const CanvasEditor = ({ templateId: propTemplateId, onSaved, hideHeader = false 
           outer={outer}
           setOuter={setOuter}
         />
-        {/* Filters (collapsible) */}
-        <div className="border-b">
-          <button
-            className="w-full text-left p-3 text-sm font-semibold"
-            onClick={() => setShowFilters((v) => !v)}
-          >
-            Filters
-          </button>
-          {showFilters && (
-            <div className="px-3 pb-3">
-              <label className="block text-xs mb-1">Course</label>
-              <select
-                className="w-full border rounded px-2 py-1 mb-2"
-                value={selectedCourse}
-                onChange={(e) => setSelectedCourse(e.target.value)}
-              >
-                <option value="">Select course</option>
-                {courses.map((c) => (
-                  <option key={c._id} value={c.Course_uuid}>
-                    {c.name}
-                  </option>
-                ))}
-              </select>
-
-              <label className="block text-xs mb-1">Batch</label>
-              <select
-                className="w-full border rounded px-2 py-1 mb-2"
-                value={selectedBatch}
-                onChange={(e) => setSelectedBatch(e.target.value)}
-              >
-                <option value="">Select batch</option>
-                {batches.map((b) => (
-                  <option key={b._id} value={b.name}>
-                    {b.name}
-                  </option>
-                ))}
-              </select>
-
-              <label className="block text-xs mb-1">Student</label>
-              <select
-                className="w-full border rounded px-2 py-1"
-                onChange={(e) => handleStudentSelect(e.target.value)}
-                value={selectedStudent?.uuid || ""}
-                disabled={bulkMode}
-              >
-                <option value="">Select a student</option>
-                {filteredStudents.map((s) => (
-                  <option key={s.uuid} value={s.uuid}>
-                    {s.firstName} {s.lastName}
-                  </option>
-                ))}
-              </select>
-
-              <div className="text-[11px] text-gray-500 mt-2">
-                Bulk mode uses the filtered list above. Use Prev/Next or swipe on mobile.
-              </div>
-            </div>
-          )}
-        </div>
+        <FiltersSection
+          showFilters={showFilters}
+          setShowFilters={setShowFilters}
+          courses={courses}
+          batches={batches}
+          selectedCourse={selectedCourse}
+          setSelectedCourse={setSelectedCourse}
+          selectedBatch={selectedBatch}
+          setSelectedBatch={setSelectedBatch}
+          handleStudentSelect={handleStudentSelect}
+          bulkMode={bulkMode}
+          filteredStudents={filteredStudents}
+          selectedStudent={selectedStudent}
+        />
 
         <FrameSection
           showFrames={showFrames}
@@ -1536,261 +1494,47 @@ const CanvasEditor = ({ templateId: propTemplateId, onSaved, hideHeader = false 
           addFrameSlot={addFrameSlot}
         />
 
-        {/* Selected object (collapsible) */}
-        {activeObj && (
-          <div className="border-b">
-            <button
-              className="w-full text-left p-3 text-sm font-semibold"
-              onClick={() => setShowSelectedSection((v) => !v)}
-            >
-              Selected Object
-            </button>
-            {showSelectedSection && (
-              <div className="px-3 pb-3">
-                <div className="flex flex-wrap gap-2">
-                  <IconButton onClick={cropImage} title="Crop"><Crop size={18} /></IconButton>
+        <SelectedObjectSection
+          activeObj={activeObj}
+          showSelectedSection={showSelectedSection}
+          setShowSelectedSection={setShowSelectedSection}
+          cropImage={cropImage}
+          canvas={canvas}
+          removeMaskAndFrame={removeMaskAndFrame}
+          setActiveObj={setActiveObj}
+          setActiveStudentPhoto={setActiveStudentPhoto}
+          saveHistory={saveHistory}
+          extractActiveImage={extractActiveImage}
+          activeStudentPhoto={activeStudentPhoto}
+          setImageZoom={setImageZoom}
+          getOverlayBox={getOverlayBox}
+        />
 
-                  <IconButton
-                    onClick={() => {
-                      const obj = activeObj;
-                      if (!obj) return;
-                      if (obj.type === "image" && obj.frameOverlay) {
-                        removeMaskAndFrame(canvas, obj, true);
-                        canvas.remove(obj);
-                      } else {
-                        canvas.remove(obj);
-                      }
-                      setActiveObj(null);
-                      setActiveStudentPhoto(null);
-                      saveHistory();
-                    }}
-                    title="Delete"
-                  >
-                    <Trash2 size={18} />
-                  </IconButton>
-
-                  <IconButton
-                    onClick={() => {
-                      const locked = !!activeObj.lockMovementX;
-                      activeObj.set({
-                        lockMovementX: !locked,
-                        lockMovementY: !locked,
-                        lockScalingX: !locked,
-                        lockScalingY: !locked,
-                        lockRotation: !locked,
-                        hasControls: locked,
-                      });
-                      canvas.renderAll();
-                    }}
-                    title="Lock/Unlock"
-                  >
-                    {activeObj?.lockMovementX ? <Unlock size={18} /> : <Lock size={18} />}
-                  </IconButton>
-
-                  {/* Extract image from frame */}
-                  {activeObj?.type === "image" && activeObj?.frameOverlay && (
-                    <IconButton onClick={extractActiveImage} title="Extract Image (remove frame)">
-                      <Move size={18} />
-                    </IconButton>
-                  )}
-                </div>
-
-                {/* Student photo quick zooms */}
-                {activeStudentPhoto && (
-                  <Stack direction="row" spacing={1} justifyContent="start" className="mt-3">
-                    <Button
-                      variant="contained"
-                      size="small"
-                      onClick={() =>
-                        setImageZoom(
-                          activeStudentPhoto,
-                          (activeStudentPhoto.scaleX || 1) * 1.1
-                        )
-                      }
-                    >
-                      Zoom In
-                    </Button>
-                    <Button
-                      variant="contained"
-                      size="small"
-                      onClick={() =>
-                        setImageZoom(
-                          activeStudentPhoto,
-                          (activeStudentPhoto.scaleX || 1) * 0.9
-                        )
-                      }
-                    >
-                      Zoom Out
-                    </Button>
-                    <Button
-                      variant="outlined"
-                      size="small"
-                      onClick={() => {
-                        const box = getOverlayBox(activeStudentPhoto);
-                        if (!box) return;
-                        const scale = Math.min(
-                          box.w / activeStudentPhoto.width,
-                          box.h / activeStudentPhoto.height
-                        );
-                        activeStudentPhoto.set({
-                          scaleX: scale,
-                          scaleY: scale,
-                          left: box.cx,
-                          top: box.cy,
-                          originX: "center",
-                          originY: "center",
-                        });
-                        canvas.requestRenderAll();
-                      }}
-                    >
-                      Reset Fit
-                    </Button>
-                  </Stack>
-                )}
-              </div>
-            )}
-          </div>
-        )}
-
-        {/* Image tools (collapsible) */}
-        {activeObj && activeObj.type === "image" && (
-          <div className="border-b">
-            <button
-              className="w-full text-left p-3 text-sm font-semibold"
-              onClick={() => setShowImageTools((v) => !v)}
-            >
-              Mask & Frame / Adjust
-            </button>
-            {showImageTools && (
-              <div className="px-3 pb-3">
-                <div className="grid grid-cols-4 gap-2 mb-3">
-                  {["rect","rounded","circle","triangle","hexagon","star","heart"].map((shape) => {
-                    const Icon = {rect:Square, rounded:Square, circle:Circle, triangle:Triangle, hexagon:Hexagon, star:Star, heart:Heart}[shape];
-                    return (
-                      <IconButton key={shape} title={shape} onClick={() => {
-                        setFrameShape(shape);
-                        applyMaskAndFrame(canvas, activeObj, shape, {
-                          stroke: frameBorder, strokeWidth: frameWidth, rx: frameCorner,
-                          absolute: adjustMode, followImage: !adjustMode
-                        });
-                        if (adjustMode) fitImageToFrame(activeObj, "cover");
-                        saveHistory();
-                      }}>
-                        <Icon size={18} />
-                      </IconButton>
-                    );
-                  })}
-                  <IconButton title="Remove Frame (keep slot)" onClick={() => { removeMaskAndFrame(canvas, activeObj, true); saveHistory(); }}>
-                    <RefreshCw size={18} />
-                  </IconButton>
-                </div>
-
-                {frameShape === "rounded" && (
-                  <div className="mb-2">
-                    <label className="block text-xs mb-1">Corner Radius</label>
-                    <Slider
-                      min={0}
-                      max={Math.floor(Math.min(activeObj?.width || 100, activeObj?.height || 100) / 2)}
-                      value={frameCorner}
-                      onChange={(_, v) => {
-                        const val = Array.isArray(v) ? v[0] : v;
-                        setFrameCorner(val);
-                        if (activeObj) {
-                          applyMaskAndFrame(canvas, activeObj, "rounded", {
-                            stroke: frameBorder, strokeWidth: frameWidth, rx: val, absolute: adjustMode, followImage: !adjustMode
-                          });
-                          if (adjustMode) fitImageToFrame(activeObj, "cover");
-                        }
-                      }}
-                    />
-                  </div>
-                )}
-
-                <div className="mb-2">
-                  <label className="block text-xs mb-1">Border Width</label>
-                  <Slider
-                    min={0}
-                    max={30}
-                    value={frameWidth}
-                    onChange={(_, v) => {
-                      const val = Array.isArray(v) ? v[0] : v;
-                      setFrameWidth(val);
-                      if (activeObj?.frameOverlay) {
-                        activeObj.frameOverlay.set({ strokeWidth: val });
-                        canvas.requestRenderAll();
-                      }
-                    }}
-                  />
-                </div>
-
-                <div className="mb-2">
-                  <label className="block text-xs mb-1">Border Color</label>
-                  <input
-                    type="color"
-                    value={frameBorder}
-                    onChange={(e) => {
-                      const col = e.target.value;
-                      setFrameBorder(col);
-                      if (activeObj?.frameOverlay) {
-                        activeObj.frameOverlay.set({ stroke: col });
-                        canvas.requestRenderAll();
-                      }
-                    }}
-                  />
-                </div>
-
-                {/* Adjust Mode */}
-                <div className="mt-4 p-3 border rounded">
-                  <div className="text-sm font-semibold mb-2">Adjust Image in Frame</div>
-                  <div className="flex flex-wrap items-center gap-2 mb-2">
-                    {!adjustMode ? (
-                      <Button size="small" variant="contained" onClick={() => enterAdjustMode(activeObj)}>
-                        Enter Adjust
-                      </Button>
-                    ) : (
-                      <Button size="small" color="secondary" variant="outlined" onClick={() => exitAdjustMode(activeObj)}>
-                        Done
-                      </Button>
-                    )}
-                    <Button size="small" variant="outlined" onClick={() => fitImageToFrame(activeObj, "contain")} disabled={!adjustMode}>
-                      Fit
-                    </Button>
-                    <Button size="small" variant="outlined" onClick={() => fitImageToFrame(activeObj, "cover")} disabled={!adjustMode}>
-                      Fill
-                    </Button>
-                    <Button size="small" variant="outlined" onClick={() => centerImageInFrame(activeObj)} disabled={!adjustMode}>
-                      Center
-                    </Button>
-                  </div>
-                  <div className="mb-2">
-                    <label className="block text-xs mb-1">Zoom</label>
-                    <Slider
-                      min={0.1}
-                      max={5}
-                      step={0.01}
-                      value={Number((activeObj?.scaleX || 1).toFixed(2))}
-                      onChange={(_, v) => setImageZoom(activeObj, Array.isArray(v) ? v[0] : v)}
-                      disabled={!adjustMode}
-                    />
-                  </div>
-                  <div className="text-[11px] text-gray-500 mt-2">
-                    Double-click an image to enter Adjust. Drag to pan under the mask. Use Fit/Fill/Center/Zoom.
-                  </div>
-                  <div className="mt-3">
-                    <Button size="small" variant="outlined" onClick={replaceActiveImage}>
-                      Replace Image
-                    </Button>
-                    {activeObj?.frameOverlay && (
-                      <Button size="small" variant="text" className="ml-2" onClick={extractActiveImage}>
-                        Extract (remove frame)
-                      </Button>
-                    )}
-                  </div>
-                </div>
-              </div>
-            )}
-          </div>
-        )}
+        <ImageToolsSection
+          activeObj={activeObj}
+          showImageTools={showImageTools}
+          setShowImageTools={setShowImageTools}
+          frameShape={frameShape}
+          setFrameShape={setFrameShape}
+          frameCorner={frameCorner}
+          setFrameCorner={setFrameCorner}
+          frameWidth={frameWidth}
+          setFrameWidth={setFrameWidth}
+          frameBorder={frameBorder}
+          setFrameBorder={setFrameBorder}
+          adjustMode={adjustMode}
+          enterAdjustMode={enterAdjustMode}
+          exitAdjustMode={exitAdjustMode}
+          fitImageToFrame={fitImageToFrame}
+          centerImageInFrame={centerImageInFrame}
+          setImageZoom={setImageZoom}
+          canvas={canvas}
+          applyMaskAndFrame={applyMaskAndFrame}
+          removeMaskAndFrame={removeMaskAndFrame}
+          replaceActiveImage={replaceActiveImage}
+          extractActiveImage={extractActiveImage}
+          saveHistory={saveHistory}
+        />
       </aside>
 
       {/* CENTER / Canva-like viewport */}
